+++ CHANGELOG.md
@@ -50,9 +50,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Extend OpenAPI document with user-generated content management endpoints [(#736)](https://github.com/wazuh/wazuh-indexer-plugins/pull/736)
 - Define Index schema for IoC content [(#744)](https://github.com/wazuh/wazuh-indexer-plugins/pull/744)
 - Add API endpoints for Logtest executions [(#746)](https://github.com/wazuh/wazuh-indexer-plugins/pull/746)
+- Add API endpoints for Rules management [(#757)](https://github.com/wazuh/wazuh-indexer-plugins/pull/757)
 - Add decoder management endpoints [(#751)](https://github.com/wazuh/wazuh-indexer-plugins/pull/751)
-- Add draft policy update endpoint [(#758)](https://github.com/wazuh/wazuh-indexer-plugins/pull/758)
 - Add kvdb management endpoints [(#752)](https://github.com/wazuh/wazuh-indexer-plugins/pull/752)
+- Add draft policy update endpoint [(#758)](https://github.com/wazuh/wazuh-indexer-plugins/pull/758) [(#762)](https://github.com/wazuh/wazuh-indexer-plugins/pull/762)
 
 ### Dependencies
 - Upgrade to Gradle 8.14.3 [(#649)](https://github.com/wazuh/wazuh-indexer-plugins/pull/649)